Inital version of html analyser
This is an attempt to see if we can read a strictly preformatted file with tags say an HTML and find out if corresponding tag of an HTML file can be identified and counted. And if yes then, them can we go inside that TAG and further find the contained attribute of that tag and count it. This will help analyse any HTML file for correctness and validate it, quickly, if it is malformed or not.

Similar concet can be used to build upon the other scripting languages too.